pyproject.toml
[project]
name = "new-socialabs-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "levenshtein>=0.27.3",
    "mpstemmer",
    "numpy>=2.2.6",
    "python-dotenv>=1.2.1",
    "sastrawi>=1.0.1",
    "scikit-learn>=1.7.2",
    "octis",
    "tensorflow>=2.20.0",
    "keras>=3.12.0",
    "networkx>=3.4.2",
    "python-louvain>=0.16",
    "pyvis>=0.3.2",
    "beanie>=2.0.1",
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
]

[dependency-groups]
dev = [
    "locust>=2.43.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[tool.uv.sources]
mpstemmer = { git = "https://github.com/ZulfiFazhar/mpstemmer.git" }
octis = { git = "https://github.com/ZulfiFazhar/OCTIS.git" }
